03.Django/02.newbie_study/requirements.txt
Django==4.0.3
psycopg2
gunicorn
supervisor
drf-yasg
djangorestframework
Pillow
